fix(desktop): keep MCP settings helpers off the client module table
The Worker pack and MCP tabs import shared settings helpers. Pulling
schemastery into that graph made the renderer miss a Host-only package
and fail plugin boot.
